Refactor task models and views: implement Task and TaskPriority classes, update imports, and remove unused code. The models folder has been removed, and the previous code inside of the task.py file has been moved to the models.py file inside of todos folder
